Moving firestore logic to a separate package
This change moves all of the firestore-specific logic out of the datastore pacakage and leaves in datastore only what is interface-level logic.

This gives us flexibility in the future to add other datastore implementations so that we can develop or run locally without a dependency on Google Cloud Platform.
