+++ backend/datastore/firestore/create_common.go
@@ -1,0 +1,22 @@
+package firestore
+
+import (
+	"context"
+	"log"
+
+	"github.com/mtlynch/whatgotdone/backend/datastore"
+)
+
+// New creates a new Datastore instance.
+func New() datastore.Datastore {
+	ctx := context.Background()
+
+	c, err := newFirestoreClient(ctx)
+	if err != nil {
+		log.Fatalln(err)
+	}
+	return &client{
+		firestoreClient: c,
+		ctx:             ctx,
+	}
+}
