+++ backend/datastore/firestore/create_dev.go
@@ -1,6 +1,6 @@
 // +build dev
 
-package datastore
+package firestore
 
 import (
 	"context"
