+++ backend/datastore/firestore/create_prod.go
@@ -1,7 +1,7 @@
 // +build !dev
 // +build !staging
 
-package datastore
+package firestore
 
 import (
 	"context"
