+++ backend/datastore/firestore/create_staging.go
@@ -1,6 +1,6 @@
 // +build staging
 
-package datastore
+package firestore
 
 import (
 	"context"
