+++ backend/datastore/firestore/entries.go
@@ -1,0 +1,43 @@
+package firestore
+
+import (
+	"strings"
+
+	"google.golang.org/api/iterator"
+
+	"github.com/mtlynch/whatgotdone/backend/types"
+)
+
+// AllEntries returns all published entries.
+func (c client) AllEntries(username string) ([]types.JournalEntry, error) {
+	entries := make([]types.JournalEntry, 0)
+	iter := c.firestoreClient.Collection(entriesRootKey).Doc(username).Collection(perUserEntriesKey).Documents(c.ctx)
+	for {
+		doc, err := iter.Next()
+		if err == iterator.Done {
+			break
+		}
+		if err != nil {
+			return nil, err
+		}
+		var j types.JournalEntry
+		doc.DataTo(&j)
+		if strings.TrimSpace(j.Markdown) == "" {
+			continue
+		}
+		entries = append(entries, j)
+	}
+	return entries, nil
+}
+
+// InsertEntry saves an entry to the datastore, overwriting any existing entry
+// with the same name and username.
+func (c client) InsertEntry(username string, j types.JournalEntry) error {
+	// Create a User document so that its children appear in Firestore console.
+	c.firestoreClient.Collection(entriesRootKey).Doc(username).Set(c.ctx, userDocument{
+		Username:     username,
+		LastModified: j.LastModified,
+	})
+	_, err := c.firestoreClient.Collection(entriesRootKey).Doc(username).Collection(perUserEntriesKey).Doc(j.Date).Set(c.ctx, j)
+	return err
+}
