+++ backend/datastore/firestore/user.go
@@ -1,0 +1,23 @@
+package firestore
+
+import (
+	"google.golang.org/api/iterator"
+)
+
+// Users returns all the users who have published entries.
+func (c client) Users() (users []string, err error) {
+	iter := c.firestoreClient.Collection(entriesRootKey).Documents(c.ctx)
+	for {
+		doc, err := iter.Next()
+		if err == iterator.Done {
+			break
+		}
+		if err != nil {
+			return nil, err
+		}
+		var u userDocument
+		doc.DataTo(&u)
+		users = append(users, u.Username)
+	}
+	return users, nil
+}
